Remove test directory if tests pass.
git-svn-id: svn://openpgp.nominet.org.uk/openpgpsdk/trunk@645 228a932f-eeec-0310-9813-fb354622df21

+++ tests/test_common.c
@@ -405,14 +405,16 @@ void cleanup()
 
     ops_crypto_finish();
 
-    return;
 
-    /* Remove test dir and files */
-    snprintf(cmd,sizeof cmd,"rm -rf %s", dir);
-    if (run(cmd))
+    if (CU_get_error()==CUE_SUCCESS)
         {
-        perror("Can't delete test directory ");
-        return;
+        /* Remove test dir and files */
+        snprintf(cmd,sizeof cmd,"rm -rf %s", dir);
+        if (run(cmd))
+            {
+            perror("Can't delete test directory ");
+            return;
+            }
         }
     }
